add team button to overview
